ENH: Rollback previous CVS commit.     Unfortunately overloaded templated member functions are not resolved correctly in     VS6 and VS70. They fail to pick the most specialized overload.

+++ Testing/Code/Common/itkVectorImageTest.cxx
@@ -53,7 +53,7 @@ int itkVectorImageTest( int, char* argv[] )
 
   const unsigned int VectorLength = 6;
   const unsigned int Dimension    = 3;
-  typedef unsigned short PixelType;
+  typedef float PixelType;
 
   {
   // Test 1.
